+++ src/typings.ts
@@ -214,7 +214,6 @@ interface Message {
 export interface PaymentRequestMessage extends Message {
   networkAddress: string
   subject?: string
-  nonce: number
   id: string
   amount: Amount
   counterParty: string
@@ -223,7 +222,6 @@ export interface PaymentRequestMessage extends Message {
 
 export interface PaymentRequestDeclineMessage {
   type: string
-  nonce: number
   id: string
   subject?: string
 }